LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_mark_exceptions-expected.txt

CONSOLE MESSAGE: line 58: ReferenceError: Can't find variable: wp_test
CONSOLE MESSAGE: line 52: ReferenceError: Can't find variable: test_true

Harness Error (FAIL), message = ReferenceError: Can't find variable: wp_test

Description

This test validates that the performance.mark() method throws a SYNTAX_ERR exception whenever a navigation timing attribute is provided for the name parameter.
